+++ src/UiPath.Caching.Abstractions/ICacheOfT.cs
@@ -5,6 +5,10 @@ public interface ICache<T>
 
     ValueTask<T?> GetAsync(CacheKey cacheKey, CancellationToken token = default);
 
+    /// <inheritdoc cref="ICache.GetAsync{T}(Span{char}, CachePolicy, CancellationToken)"/>
+    ValueTask<T?> GetAsync(Span<char> cacheKey, CancellationToken token = default) =>
+        GetAsync(new CacheKey(cacheKey), token);
+
     ValueTask<KeyValuePair<CacheKey, T?>[]> GetAsync(CacheKey[] cacheKeys, CancellationToken token = default);
 
     ValueTask<T?> GetOrAddAsync(CacheKey cacheKey, Func<CancellationToken, Task<T?>> generator, CancellationToken token = default);